% Generated by roxygen2 (4.0.0): do not edit by hand
\name{spoon}
\alias{spoon}
\title{A Fake Function}
\usage{
spoon(spoon = FALSE)
}
\arguments{
\item{spoon_or_not}{whether or not there is a spoon. Default is FALSE}
}
\description{
This function isn't real
}
\examples{
spoon()
}
\keyword{spoons}

